sync

+++ mgamdata/mm/mgam_models.py
@@ -248,7 +248,7 @@ class mgam_Seg2D_Lite(mgam_Seg_Lite):
         
         preds = torch.zeros(
             size=(batch_size, self.num_classes, h_img, w_img),
-            dtype=torch.float16,
+            dtype=torch.float32,
             device=accumulate_device
         )
         count_mat = torch.zeros(
